Move useTabBarHook function to a separate hook file
